fix: donation amount type from int to double

+++ Server/SportReserve_Shared/Models/Reservation/Add/AddAnimalShelterRace.cs
@@ -6,7 +6,7 @@ namespace SportReserve_Shared.Models.Reservation.Add
     public class AddAnimalShelterRace : ReservationBase
     {
         public string? DogSize { get; set; }
-        public int? DonationAmount { get; set; }
+        public double? DonationAmount { get; set; }
 
         [Required]
         public string EmergencyContact { get; set; } = string.Empty;
